facturacion/requirements.txt
# Para conexión a MariaDB
pymysql
python-dotenv
requests
flask
flask_sqlalchemy
pymysql
pika
python-dotenv
flask-sqlalchemy